pyproject.toml
[project]
name = "model-training"
version = "0.1.0"
description = "Classify training data using OpenAI Batch API"
authors = [
    {name = "Euan"}
]
readme = "README.md"
requires-python = "^3.12"
dependencies = [
    "datasets (>=4.5.0,<5.0.0)",
    "huggingface-hub (>=1.4.1,<2.0.0)",
    "openai (>=2.21.0,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "transformers (>=5.1.0,<6.0.0)",
    "scikit-learn (>=1.8.0,<2.0.0)",
    "accelerate (>=1.12.0,<2.0.0)",
    "torch (>=2.10.0,<3.0.0) ; python_version >= \"3.12\" and python_version < \"3.15\""
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
